+++ packages/providers/lib.esm/etherscan-provider.js
@@ -312,20 +312,20 @@ export class EtherscanProvider extends BaseProvider {
                     url += apiKey;
                     const logs = yield get(url, null, getResult);
                     // Cache txHash => blockHash
-                    let txs = {};
+                    let blocks = {};
                     // Add any missing blockHash to the logs
                     for (let i = 0; i < logs.length; i++) {
                         const log = logs[i];
                         if (log.blockHash != null) {
                             continue;
                         }
-                        if (txs[log.transactionHash] == null) {
-                            const tx = yield this.getTransaction(log.transactionHash);
-                            if (tx) {
-                                txs[log.transactionHash] = tx.blockHash;
+                        if (blocks[log.blockNumber] == null) {
+                            const block = yield this.getBlock(log.blockNumber);
+                            if (block) {
+                                blocks[log.blockNumber] = block.hash;
                             }
                         }
-                        log.blockHash = txs[log.transactionHash];
+                        log.blockHash = blocks[log.blockNumber];
                     }
                     return logs;
                 }