+++ packages/providers/lib.esm/formatter.js
@@ -115,6 +115,9 @@ export class Formatter {
     // Requires a BigNumberish that is within the IEEE754 safe integer range; returns a number
     // Strict! Used on input.
     number(number) {
+        if (number === "0x") {
+            return 0;
+        }
         return BigNumber.from(number).toNumber();
     }
     // Strict! Used on input.