+++ packages/providers/lib/etherscan-provider.js
@@ -210,7 +210,7 @@ var EtherscanProvider = /** @class */ (function (_super) {
     };
     EtherscanProvider.prototype.perform = function (method, params) {
         return __awaiter(this, void 0, void 0, function () {
-            var url, apiKey, get, _a, postData, error_1, postData, error_2, topic0, logs, txs, i, log, tx, _b;
+            var url, apiKey, get, _a, postData, error_1, postData, error_2, topic0, logs, blocks, i, log, block, _b;
             var _this = this;
             return __generator(this, function (_c) {
                 switch (_c.label) {
@@ -389,7 +389,7 @@ var EtherscanProvider = /** @class */ (function (_super) {
                         return [4 /*yield*/, get(url, null, getResult)];
                     case 20:
                         logs = _c.sent();
-                        txs = {};
+                        blocks = {};
                         i = 0;
                         _c.label = 21;
                     case 21:
@@ -398,16 +398,16 @@ var EtherscanProvider = /** @class */ (function (_super) {
                         if (log.blockHash != null) {
                             return [3 /*break*/, 24];
                         }
-                        if (!(txs[log.transactionHash] == null)) return [3 /*break*/, 23];
-                        return [4 /*yield*/, this.getTransaction(log.transactionHash)];
+                        if (!(blocks[log.blockNumber] == null)) return [3 /*break*/, 23];
+                        return [4 /*yield*/, this.getBlock(log.blockNumber)];
                     case 22:
-                        tx = _c.sent();
-                        if (tx) {
-                            txs[log.transactionHash] = tx.blockHash;
+                        block = _c.sent();
+                        if (block) {
+                            blocks[log.blockNumber] = block.hash;
                         }
                         _c.label = 23;
                     case 23:
-                        log.blockHash = txs[log.transactionHash];
+                        log.blockHash = blocks[log.blockNumber];
                         _c.label = 24;
                     case 24:
                         i++;